refactor(e2e): clean up scale test, remove dead code, extract constants
- Merge pprof_hook.go into main_test.go (single test infra file)
- Delete unused cr_client.go and grove_config_k8s.go (duplicate utils)
- Extract magic strings/ints to named constants
- Fix parsedSelector hot-path mutation in condition/pod.go
- Add MkdirAll for output directory, use test-named subdirectory
