--- operator/e2e/utils/cr_client.go
@@ -1,46 +1,0 @@
-// /*
-// Copyright 2026 The Grove Authors.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-// */
-
-package utils
-
-import (
-	"fmt"
-
-	corev1alpha1 "github.com/ai-dynamo/grove/operator/api/core/v1alpha1"
-	"k8s.io/apimachinery/pkg/runtime"
-	clientgoscheme "k8s.io/client-go/kubernetes/scheme"
-	"k8s.io/client-go/rest"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-)
-
-// NewCRClient creates a controller-runtime client with the Grove scheme registered.
-// This is needed for conditions that use typed CR access (e.g. PCSDeletedCondition).
-func NewCRClient(restConfig *rest.Config) (client.Client, error) {
-	scheme := runtime.NewScheme()
-	if err := clientgoscheme.AddToScheme(scheme); err != nil {
-		return nil, fmt.Errorf("add client-go scheme: %w", err)
-	}
-	if err := corev1alpha1.AddToScheme(scheme); err != nil {
-		return nil, fmt.Errorf("add Grove scheme: %w", err)
-	}
-
-	cl, err := client.New(restConfig, client.Options{Scheme: scheme})
-	if err != nil {
-		return nil, fmt.Errorf("create controller-runtime client: %w", err)
-	}
-
-	return cl, nil
-}
